Update navbar drawer icons
