--- src/pages/About.js
@@ -1,27 +1,0 @@
-import React from "react";
-
-const About = () => (
-  <div id="about" className="section">
-    <h1>This is about page!</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-    <h1>About Page</h1>
-  </div>
-);
-
-export default About;
