--- src/pages/Education.js
@@ -1,38 +1,0 @@
-import React from "react";
-
-const Education = () => (
-  <div id="education" className="section">
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-    <h1>Education Page</h1>
-  </div>
-);
-
-export default Education;
