--- src/pages/Projects.js
@@ -1,32 +1,0 @@
-import React from "react";
-
-const Projects = () => (
-  <div id="projects" className="section">
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-    <h1>Projects Page</h1>
-  </div>
-);
-
-export default Projects;
