--- src/pages/Work.js
@@ -1,38 +1,0 @@
-import React from "react";
-
-const Work = () => (
-  <div id="work" className="section">
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-    <h1>Work Page</h1>
-  </div>
-);
-
-export default Work;
