+++ src/sections/About.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => (
+  <div id="about" className="section">
+    <p> Hi, I'm</p>
+    <h1>Kevin</h1>
+    <p>A software engineer</p>
+  </div>
+);
+
+export default About;
