+++ src/sections/Education.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Education = () => (
+  <div id="education" className="section">
+    <h1>Education:</h1>
+  </div>
+);
+
+export default Education;
