+++ src/sections/Projects.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Projects = () => (
+  <div id="projects" className="section">
+    <h1>Projects:</h1>
+  </div>
+);
+
+export default Projects;
