+++ src/sections/Work.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Work = () => (
+  <div id="work" className="section">
+    <h1>Work History:</h1>
+  </div>
+);
+
+export default Work;
