+++ src/sections/index.js
@@ -1,0 +1,9 @@
+import About from "./About";
+import Projects from "./Projects";
+import Work from "./Work";
+import Education from "./Education";
+
+export { About };
+export { Projects };
+export { Work };
+export { Education };
